+++ test/unit/fuzz-tools.test.js
@@ -279,6 +279,7 @@ import * as cosyntropinV299 from '../../lib/cosyntropin-v299.js';
 import * as avFistulaV300 from '../../lib/av-fistula-v300.js';
 import * as drSeverityV301 from '../../lib/dr-severity-v301.js';
 import * as isisV302 from '../../lib/isis-v302.js';
+import * as anaphylaxisV303 from '../../lib/anaphylaxis-v303.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -503,6 +504,7 @@ const MODULES = {
   'av-fistula-v300.js': avFistulaV300,
   'dr-severity-v301.js': drSeverityV301,
   'isis-v302.js': isisV302,
+  'anaphylaxis-v303.js': anaphylaxisV303,
   'tb-testing.js': tbTesting,
 };